requirements.txt
asyncinotify
pyyaml
